// src/core/types.ts
export interface PostMeta {
  title: string
  slug: string
  subtitle?: string
  description?: string
  image?: string
  cover?: string
  publishedAt: string
  updatedAt?: string
}

export interface SiteMeta {
  title: string
  description: string
  siteUrl: string
  image: string
  publisher: string
}